Allow a schema.*Resource to give its own aliases
Note: This is a differnt concept of alias then a schema alias. This
alias is used to rewrite the schema, pointing references of the alias to
the original TK.

Use case:

This allows a component resource to embed the state of a custom resource
and get the expected result (instead of a seperate
`#/type/pkg:mod:FooState`).