Fix a few issues with diff
The diff routine has three issues:
1. It does not correctly handled changes in nested proeprties
2. It inverts adds/deletes in the diffs it reports, leading to confusing diff presentation
3. It is untested.

This PR fixes these issues.

The first issue was introduced by the use of `ignoreKeys` callbacks on PropertyValue Diff.  Unfortunately, that feature in pulumi/pulumi is not really safe to use, because it doesn't return full property paths, only the final key name in the nested path.  So the caller cannot determine which nested key is really being referenced.  The only other place it is used in pulumi/pulumi happens to only need to be correct for top-level properties, so happens to work.  But it does not work for the way it was being used here.

The second issue is an inversion of the order of arguments to PropertyMap#Diff.

Test coverage for all the core of Diff is added.  Additional coverage for ignoreChanges, CustomDiff, and Replacements should also be added in the future.