Hook up github.com/cloudy-sky-software/pulschema to `infer`
This commit adds a package infer/openapipulschema that uses pulschema to
load an API spec and create a Pulumi schema from it. A new provider
is created based on the schema plus provided metadata.

Theoretically, this would allow generating the provider and schema parts
for an OpenAPI-based provider, leaving only the CRUD implementations.
However, I wasn't able to successfully load an OpenAPI spec with
pulschema so far.