[Breaking] Remove p.Context in favor of context.Context
This simplifies the API of a provider and makes it much easier to go back and forth
between code designed to work with pulumi-go-provider and other Go code.

Migration guide:

- Replace all instances of `"github.com/pulumi/pulumi-go-provider".Context` with
  `context.Context`.
- Methods on a `ctx "github.com/pulumi/pulumi-go-provider".Context` have become
  free-standing functions on `"github.com/pulumi/pulumi-go-provider"`
  - `ctx.RuntimeInformation` has become
  `"github.com/pulumi/pulumi-go-provider".GetRunInfo(ctx)`.
  - `ctx.Log*` has been replaced by
  `"github.com/pulumi/pulumi-go-provider".GetLogger(ctx).*`.
  - `"github.com/pulumi/pulumi-go-provider/infer".CtxFromPulumiContext(ctx)` has been
    replaced by `ctx.Context()`.

I'm sorry to break existing users but I think this design is strictly better in the long
term. Go has standardized on a `context.Context` type and the more Go's ecosystem gets
behind it the easier everyone's life will be.