Introduce AssetOrArchive to support the SDK's Asset type which can be both (#242)
Resolves #237.

This PR addresses an issue with the design of the [assets and archives
support](https://www.pulumi.com/docs/concepts/assets-archives/) in
Pulumi core. Quoting #237:

> When defining a property of type resource.Asset, the generated schema
will contain a $ref of pulumi.json#/Asset. However, in the SDKs, an
"asset" is represented by pulumi.AssetOrArchive because an archive is
currently always a valid asset.

> Therefore, when deserializing the PropertyValue for the provider to
use, it's valid for the user to have specified one of the Archive types
which then isn't deserializable into the resource.Asset field within the
provider.

To work around this, we define a new type `types.AssetOrArchive` here in
pulumi-go-provider. In the schema, it's represented as a standard
`pulumi.json#/Asset` for interoperability. An incoming core `Asset` is
translated into a `types.AssetOrArchive`, in `Decode`, with either Asset
or Archive populated. Vice versa, `Encode` pulls the asset/archive out
of AssetOrArchive.

## Notes
- The decode and encode implementations are bespoke and there might be
opportunities to move it to a new, more generic extension point.

---------

Co-authored-by: Ian Wahbe <ian@wahbe.com>