Change the form of `infer.Component`
Instead of taking a `[Anchor, Args, State]` triple like custom resources, we change the
shape to directly pass the `New*` construct function.

Existing users of `infer.Component` will need to make the changes seen in our example code.

The motivation for this change is to make the API of "packages" components (consumable
from any language) compatible with the API that we suggest for "local"
components (consumable from other Go programs). After this change, in both cases a
component would require two parts:

```
type RandomLogin struct { ... }

func NewRandomLogin(ctx *pulumi.Context, name string, args RandomLoginArgs, opts ...pulumi.ResourceOption) { ... }
```

so, a type and a corresponding constructor function.

This aligns well with the story that we are building towards for all languages, and thus
justifies the breaking change. We believe that a combination of a preview 0.x version of
the library, low number of end users, and simplicity of migration warrants the breaking
change to get clean API in the future.