Migrate `resource.PropertyValue` to `property.Value` in public APIs (#344)
We believe that `property.Value` is a safer and less mistake prone API
compared to `resource.PropertyValue`. While we break all the things,
let's take the opportunity to use the better representation of pulumi
values. Critically, `property.Value` is an immutable data type. **We
require an immutable value representation to handle a native set type**,
which we would like to add in the future.

This PR changes all public APIs in this repo to use `property.Value` and
`property.Map` instead of `resource.PropertyValue` and
`resource.PropertyMap`. Internal values were migrated only as
convenient, primarily to reduce the size of this already large PR.

Fixes https://github.com/pulumi/pulumi-go-provider/issues/347

---------

Co-authored-by: Eron Wright <eron@pulumi.com>