Rename Olds/News -> State/Inputs (#352)
Olds/News is a Pulumi naming convention that isn't immediately obvious
to newcomers.

We're already using State/Inputs terminology, which is much more
obvious. Let's rename these for consistency.

Edit: We could also go the other way and rename State/Inputs to →
Olds/News if we want to keep things as close as possible to the
protocol.