Remove `InputDependencies` from Provider interface (#350)
This PR removes the formal "dependencies" map that often accompanies
property maps on Provider requests and responses. Any dependences on an
incoming RPC request are simply folded into their respective properties.
The resultant RPC response never contains a dependencies map, and the
provider asserts that the caller MUST accept output values.

The request/response structs are changing like this:

![image](https://github.com/user-attachments/assets/852db98f-c6c3-4aa6-b97f-42bce63f632b)


The middleware.rpc package takes pains to downgrade input dependencies
based on the wrapped provider's capabilities (Call and Construct ONLY).
The wrapped provider MAY return a dependencies map in all modes.

**To discuss**: we know that dependency maps are often generated by
walking an output value (recursively). When we use upgrade logic to
re-attach the dependencies to the property values, does that cause some
"garbage" to build up on those values? Should the caller provide a
dependency map only in a downgrade situation? When may it be ignored?
Should we filter out the deps that are present on the children?

Here, for example, we re-attach the dependencies map to the property
values. If those deps were extracted from deep children, now they're
attached directly to the top-level value. Very concerning to me.

https://github.com/pulumi/pulumi-go-provider/pull/350/files#diff-1bb2cdf24b7aa1b1986f79d4ea823ec921a95ba94f6436119cb2a74c21ebff82R876-R883

on-wire:
```plain
v
├── k1 (urn1)
└── k2
    └── k3 (urn2)
```

after upgrade:
```plain
v (urn1, urn2)
├── k1 (urn1)
└── k2
    └── k3 (urn2)
```

My best idea here is to re-attach ONLY the deps that aren't present on
the children. WDYT?

Update: It is assumed that, for a given value, the dependencies are
either within the value itself OR in the dependency map. So, if the
value has any dependencies, you needn't incorporate the map, and doing
so would have the erroneous effect of folding the child dependencies
into the parent.

Closes #346