Split up BuildAndRun (#364)
It's very common (especially in testing) to construct a provider and
then pass it around before acting on it.

https://github.com/pulumi/pulumi-docker-build/blob/30a01b689388605b53cf10594ad593c9f270cf15/provider/internal/provider.go#L68

https://github.com/pulumi/pulumi-provider-boilerplate/blob/5c43e6c38e62b35b3352dc535f6e83c6cf4ab32f/provider/provider.go#L28

This means `BuildAndRun` is too restrictive, so this PR breaks out
`Build` into a standalone method which returns an error if validation
fails.

`Run` is added to go-provider's `Provider` struct to spare the user the
unintuitive `provider.Main` call. Name and Version make this a little
kludgy because we don't currently track those on the `Provider` itself.