Add WithProvider and WithProviderF server options (#370)
This consolidates all of the NewServer methods, and adds WithProvider
and WithProviderF options. This way servers can continue to be created
with concrete or lazy providers.

A `WithVersion` would make sense IMO, but we can do that later.

(It's odd that we use an option pattern here and a builder pattern for
the provider.)