Keep internal metadata under __internal (#377)
The `__internal` object is automatically excluded from diffs by the
engine as of https://github.com/pulumi/pulumi/pull/18440.

Bumping pu/pu to pull in that change resolves #288 for `__internal`
properties, but the issue still exists for the additional metadata
introduced in #298.

Rather than special-case the new metadata, let's move it into the
internal object.