Don't replace providers on version changes (#410)
Followup to https://github.com/pulumi/pulumi-go-provider/pull/407 and
https://github.com/pulumi/pulumi/pull/20401.

If a provider doesn't define `DiffConfig` we currently replace it on any
changes (https://github.com/pulumi/pulumi-go-provider/issues/409).

Changing the default behavior to instead update is a breaking change in
behavior, so we can't do that.

However, the version field in particular is currently broken -- changing
the version and nothing else should _not_ trigger a replacement.
Special-casing this field to only cause updates is an appropriate bug
fix.